split ryu specials into folder
