+++ fighters/ryu/src/acmd/specials/mod.rs
@@ -1,0 +1,12 @@
+use super::*;
+mod special_hi;
+mod special_lw;
+mod special_n;
+mod special_s;
+
+pub fn install() {
+    special_hi::install();
+    special_lw::install();
+    special_n::install();
+    special_s::install();
+}
