+++ fighters/ryu/src/acmd/specials/special_lw.rs
@@ -1,0 +1,7 @@
+
+use super::*;
+
+pub fn install() {
+    // smashline::Agent::new("ryu")
+    //     .install();
+}
